Move related Component to kamiza directory
